+++ src/UIComponents/Timeline/Timeline.js
@@ -1,16 +1,20 @@
 import React from "react";
+import PropTypes from "prop-types";
 import className from "classnames";
 import "./Timeline.scss";
 
-const Timeline = ({ items, renderItem, view }) => {
+const Timeline = ({ items, renderItem, customStopItem }) => {
   const classList = className({
     timeline: true,
-    single: view === "SINGLE",
   });
+
   return (
     <div className={classList}>
       {items.map((item) => (
         <div key={item.id} className="timeline-block">
+          <div className="timeline-block-stops">
+            {customStopItem ? customStopItem() : null}
+          </div>
           {renderItem(item)}
         </div>
       ))}
@@ -18,8 +22,10 @@ const Timeline = ({ items, renderItem, view }) => {
   );
 };
 
-Timeline.defaultProps = {
-  view: "SINGLE",
+Timeline.defaultProps = {};
+
+Timeline.propTypes = {
+  customStopItem: PropTypes.func,
 };
 
 export default Timeline;