+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ### v0.1.0
+
+- Timeline new structure